requirements.txt
streamlit>=1.31.0
requests>=2.31.0
python-dotenv>=1.0.1
